+readme

+++ language/great_grammar.cpp
@@ -256,7 +256,6 @@ ma_ty get_E (parsed_line_reader* pl_reader)
     return val;
     }
 
-
 /*!
 @brief Parse multiply and division;
        Descend to power etc;
@@ -303,7 +302,6 @@ ma_ty get_S (parsed_line_reader* pl_reader)
 
     return val;
     }
-
 
 /*!
 @brief Parse function calls;
@@ -361,7 +359,6 @@ ma_ty get_F (parsed_line_reader* pl_reader)
         }
     }
 
-
 /*!
 @brief [P]rimary level: parse parenthesized expressions;
        Descend to values and variables parse;
@@ -392,7 +389,7 @@ ma_ty get_N (parsed_line_reader* pl_reader)
     {
     if (TOKEN_TYPE == T_VAL)
         {
-        int val = 0;
+        double val = 0;
 
         val = gfs(pl[TOKEN_ID++].content.val);
